Broke out some code to utility files
