+++ client/components/TestResults.jsx
@@ -13,8 +13,6 @@ import Fab from '@material-ui/core/Fab';
 import Snackbar from '@material-ui/core/Snackbar';
 import AddIcon from '@material-ui/icons/Add';
 
-import moment from 'moment';
-
 import {
   subscribeToPipelineNames,
   subscribeToTestResultUpdates,
@@ -27,6 +25,7 @@ import {
   emitTestResultRemove
 } from "../api";
 
+import { sortReports, convertReport } from './TestResults.util';
 import TestReport from './TestReport';
 import AddTest from './AddTest';
 
@@ -38,9 +37,6 @@ const createStyles = darkTheme => ({
     color: darkTheme ? '#000' : '#fff'
   }
 });
-
-// From latest report to this number of days back in time
-const daysInterval = 20;
 
 export default class TestResults extends React.Component {
 
@@ -67,7 +63,7 @@ export default class TestResults extends React.Component {
   }
   testResultsListener = (testReports) => {
     this.setState({
-      testReports: testReports.map(this.convertReport).sort(this.sortReports)
+      testReports: testReports.map(convertReport).sort(sortReports)
     });
   }
   testMessageListener = (message) => {
@@ -104,80 +100,6 @@ export default class TestResults extends React.Component {
     this.setState({
       addTestDialogOpened: true
     });
-  }
-
-  /**
-   * Sort reports. Failed first, then time for latest report
-   */
-  sortReports(r1, r2) {
-    const latest1 = r1.history[r1.history.length - 1];
-    const latest2 = r2.history[r2.history.length - 1];
-    if (latest2.failed > 0 && latest1.failed <= 0) {
-      return 1;
-    }
-    if (latest1.failed > 0 && latest2.failed <= 0) {
-      return -1;
-    }
-    if (latest2.when > latest1.when) {
-      return 1;
-    }
-    return -1;
-  }
-
-  /**
-   * Converts report data to report object
-   */
-  convertReport(report) {
-    // Report model
-    const reportView = {
-      id: report._id,
-      title: `${report.pipeline} (${report.stage})`,
-      subtitle: report.job
-    };
-    if (report.cucumber) {
-      // Create chart history data      
-      reportView.history = report.cucumber
-        // Sort by time ascending
-        .sort((a, b) => {
-          return a.timestamp > b.timestamp ? 1 : -1;
-        })
-        // Filter reports that are not in defined interval
-        .filter((report, idx, arr) => {
-          // Latest test case = last in list
-          const latestTestTime = moment(arr[arr.length - 1].timestamp);
-          const currTestTime = moment(report.timestamp);
-          return latestTestTime.diff(currTestTime, 'days') <= daysInterval;
-        })
-        .reduce((acc, c) => {
-          const errors = [];
-          let passed = 0;
-          let failed = 0;
-          c.features.forEach((feature) => {
-            feature.scenarios.forEach((scenario) => {
-              scenario.steps.forEach((step) => {
-                if (step.result === 'passed') {
-                  passed++;
-                } else {
-                  failed++;
-                  errors.push({
-                    test: scenario.name,
-                    message: step.error,
-                  });
-                }
-              })
-            })
-          })
-          acc.push({
-            passed: passed,
-            failed: failed,
-            errors: errors,
-            when: c.timestamp
-          });
-          return acc;
-        }, []);
-
-    }
-    return reportView;
   }
 
   resetMessage() {
